refactor(evaluatorq-py): move OrqResponsesTarget to openresponses/target.py, remove redteam->simulation edge (RES-897)
