+++ packages/evaluatorq-py/tests/redteam/test_orq_responses_target_as_agent_target.py
@@ -13,7 +13,7 @@ from unittest.mock import AsyncMock, MagicMock
 import pytest
 
 from evaluatorq.contracts import AgentResponse, AgentTarget, LLMCallConfig, Message
-from evaluatorq.simulation.target import OrqResponsesTarget
+from evaluatorq.openresponses.target import OrqResponsesTarget
 
 
 def _make_client() -> MagicMock:
